+++ ctor/tests/expand-darwin/ctor_priority.expanded.rs
@@ -7,13 +7,16 @@ fn foo() {
         };
     }
     const _: () = {
-        #[allow(unused_unsafe)]
-        fn __ctor_private() {
+        #[allow(unsafe_code, unused_unsafe)]
+        extern "C" fn __ctor_private() {
             { { __ctor_private_inner() } }
         }
-        #[link_section = "__DATA,CTOR,regular,no_dead_strip"]
+        #[link_section = "__DATA,_CTOR0_ISIZE_FN,regular,no_dead_strip"]
         #[used]
-        static __CTOR_ENTRY: (fn(), u16) = (__ctor_private, 1);
+        pub static CTOR: ::ctor::collect::Constructor = ::ctor::collect::Constructor {
+            priority: 1,
+            ctor: __ctor_private,
+        };
     };
     { __ctor_private_inner() }
 }