+++ ctor/tests/expand-darwin/ctor_static.expanded.rs
@@ -10,14 +10,14 @@ static STATIC_CTOR: ::ctor::statics::Static<HashMap<u32, &'static str>> = {
     unsafe { ::ctor::statics::Static::<HashMap<u32, &'static str>>::new(init) }
 };
 const _: () = {
-    #[allow(unsafe_code)]
-    #[link_section = "__DATA,__mod_init_func,mod_init_funcs"]
+    #[allow(unsafe_code, unused_unsafe)]
+    extern "C" fn __ctor_private() {
+        { _ = &*STATIC_CTOR }
+    }
+    #[link_section = "__DATA,_CTOR0_ISIZE_FN,regular,no_dead_strip"]
     #[used]
-    static __CTOR_PRIVATE_REF: unsafe extern "C" fn() = {
-        #[allow(unused_unsafe)]
-        extern "C" fn __ctor_private() {
-            { _ = &*STATIC_CTOR }
-        }
-        __ctor_private
+    pub static CTOR: ::ctor::collect::Constructor = ::ctor::collect::Constructor {
+        priority: 0,
+        ctor: __ctor_private,
     };
 };