+++ ctor/tests/expand-linux/ctor_anon.expanded.rs
@@ -9,7 +9,7 @@ const _: () = {
         }
         const _: () = {
             #[allow(unsafe_code)]
-            #[link_section = ".init_array"]
+            #[link_section = ".init_array.000"]
             #[used]
             static __CTOR_PRIVATE_REF: unsafe extern "C" fn() = {
                 #[allow(unused_unsafe)]
@@ -32,7 +32,7 @@ const _: () = {
         }
         const _: () = {
             #[allow(unsafe_code)]
-            #[link_section = ".init_array"]
+            #[link_section = ".init_array.000"]
             #[used]
             static __CTOR_PRIVATE_REF: unsafe extern "C" fn() = {
                 #[allow(unused_unsafe)]