+++ ctor/tests/expand-linux/ctor_priority_late.expanded.rs
@@ -1,24 +1,22 @@
 use ctor::ctor;
-/// Doc 1
-/// Doc 2
 #[allow(dead_code)]
-unsafe fn foo() {
-    unsafe fn __ctor_private_inner() {
+fn foo() {
+    fn __ctor_private_inner() {
         {
             ::std::io::_print(format_args!("foo\n"));
         };
     }
     const _: () = {
         #[allow(unsafe_code)]
-        #[link_section = ".init_array.000"]
+        #[link_section = ".init_array.65535"]
         #[used]
         static __CTOR_PRIVATE_REF: unsafe extern "C" fn() = {
             #[allow(unused_unsafe)]
             extern "C" fn __ctor_private() {
-                { unsafe { __ctor_private_inner() } }
+                { { __ctor_private_inner() } }
             }
             __ctor_private
         };
     };
-    unsafe { __ctor_private_inner() }
+    { __ctor_private_inner() }
 }