+++ ctor/tests/expand-linux/ctor_priority_late.rs
@@ -1,0 +1,6 @@
+use ctor::ctor;
+
+#[ctor(priority = late)]
+fn foo() {
+    println!("foo");
+}