+++ ctor/tests/target-test/ctor_priority.rs
@@ -1,0 +1,21 @@
+use ctor::declarative::ctor;
+
+ctor!(
+    #[ctor(unsafe, priority = early)]
+    fn early() {}
+);
+
+ctor!(
+    #[ctor(unsafe, priority = 1)]
+    fn priority1() {}
+);
+
+ctor!(
+    #[ctor(unsafe, priority = late)]
+    fn late() {}
+);
+
+ctor!(
+    #[ctor(unsafe, priority = naked)]
+    fn naked() {}
+);